+++ src/main/kotlin/com/calmwolfs/bedwar/data/game/OtherInventoryData.kt
@@ -13,7 +13,7 @@ import net.minecraft.network.play.server.S2FPacketSetSlot
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 
 object OtherInventoryData {
-    var currentInventory: Inventory? = null
+    private var currentInventory: Inventory? = null
     private var acceptItems = false
     private var lateEvent: InventoryUpdatedEvent? = null