+++ src/main/kotlin/com/calmwolfs/bedwar/features/party/PartyGameStats.kt
@@ -10,40 +10,40 @@ import kotlinx.coroutines.launch
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
 import kotlin.time.Duration.Companion.milliseconds
 
-class PartyGameStats {
+object PartyGameStats {
     // todo deal with nicks
     private val config get() = BedWarMod.feature.party.matchStats
-    private var gamePartyMembers = mutableMapOf<String, BedwarsGameStat>()
+    var gameMembers = mutableMapOf<String, BedwarsGameStat>()
 
     @SubscribeEvent
     fun onGameStart(event: StartGameEvent) {
-        gamePartyMembers.clear()
+        gameMembers.clear()
         for (member in PartyUtils.partyMembers) {
-            gamePartyMembers[member] = BedwarsGameStat(0, 0, 0)
+            gameMembers[member] = BedwarsGameStat(0, 0, 0)
         }
-        gamePartyMembers[HypixelUtils.currentName] = BedwarsGameStat(0, 0, 0)
+        gameMembers[HypixelUtils.currentName] = BedwarsGameStat(0, 0, 0)
     }
 
     @SubscribeEvent
     fun onKill(event: KillEvent) {
-        if (event.killer in gamePartyMembers) {
-            val playerStats = gamePartyMembers[event.killer] ?: BedwarsGameStat(0, 0, 0)
+        if (event.killer in gameMembers) {
+            val playerStats = gameMembers[event.killer] ?: BedwarsGameStat(0, 0, 0)
             playerStats.kills = playerStats.kills + 1
         }
     }
 
     @SubscribeEvent
     fun onFinal(event: FinalKillEvent) {
-        if (event.killer in gamePartyMembers) {
-            val playerStats = gamePartyMembers[event.killer] ?: BedwarsGameStat(0, 0, 0)
+        if (event.killer in gameMembers) {
+            val playerStats = gameMembers[event.killer] ?: BedwarsGameStat(0, 0, 0)
             playerStats.finals = playerStats.finals + 1
         }
     }
 
     @SubscribeEvent
     fun onBed(event: BedBreakEvent) {
-        if (event.player in gamePartyMembers) {
-            val playerStats = gamePartyMembers[event.player] ?: BedwarsGameStat(0, 0, 0)
+        if (event.player in gameMembers) {
+            val playerStats = gameMembers[event.player] ?: BedwarsGameStat(0, 0, 0)
             playerStats.beds = playerStats.beds + 1
         }
     }
@@ -51,16 +51,14 @@ class PartyGameStats {
     @SubscribeEvent
     fun onGameEnd(event: EndGameEvent) {
         if (!config.enabled) return
-        if (gamePartyMembers.size < 2 && !config.showSolo) return
         val winner = event.winningTeam == BedwarsUtils.currentTeamName
         BedWarMod.coroutineScope.launch {
-            delay(1250.milliseconds)
+            delay(config.statsDelay.milliseconds)
             ChatUtils.chat("")
             ChatUtils.chat("§6[BedWar] §7Parties match stats")
             var line = ""
-            for ((player, stats) in gamePartyMembers) {
+            for ((player, stats) in gameMembers) {
                 val isPlayer = HypixelUtils.currentName == player
-                if (isPlayer && !config.showSolo) continue
 
                 val kills = StringUtils.optionalPlural(stats.kills, "§7Kill", "§7Kills")
                 val finals = StringUtils.optionalPlural(stats.finals, "§7Final Kill", "§7Final Kills")
@@ -73,16 +71,13 @@ class PartyGameStats {
                     line = "${stats.kills} ${stats.finals} ${stats.beds}"
                     if (config.actionType == 1 || config.actionType == 3) {
                         ClipboardUtils.copyToClipboard(line)
-                        if (config.actionType == 1) {
-                            ChatUtils.chat("§a[BedWar] §7Copied match stats to clipboard!")
-                        }
                     }
                 }
             }
             ChatUtils.chat("")
             if (!winner && !config.sendOnLoss) return@launch
             if (line != "" && (config.actionType == 2 || config.actionType == 3) && PartyUtils.partyMembers.isNotEmpty()) {
-                delay(250.milliseconds)
+                delay(config.sendDelay.milliseconds)
                 ChatUtils.sendCommandToServer("pc $line")
             }
         }