+++ src/main/kotlin/com/calmwolfs/bedwar/utils/ItemUtils.kt
@@ -1,6 +1,8 @@
 package com.calmwolfs.bedwar.utils
 
+import net.minecraft.init.Items
 import net.minecraft.item.ItemStack
+import net.minecraftforge.common.util.Constants
 
 object ItemUtils {
     fun ItemStack.getLore(): List<String> {
@@ -18,4 +20,13 @@ object ItemUtils {
         set(value) {
             setStackDisplayName(value)
         }
+
+    fun ItemStack.getSkullTexture(): String? {
+        if (item != Items.skull) return null
+        if (tagCompound == null) return null
+        val nbt = tagCompound
+        if (!nbt.hasKey("SkullOwner")) return null
+        return nbt.getCompoundTag("SkullOwner").getCompoundTag("Properties")
+            .getTagList("textures", Constants.NBT.TAG_COMPOUND).getCompoundTagAt(0).getString("Value")
+    }
 }