+++ src/main/kotlin/com/calmwolfs/bedwar/utils/LocationUtils.kt
@@ -1,0 +1,8 @@
+package com.calmwolfs.bedwar.utils
+
+import com.calmwolfs.bedwar.data.types.getModVector
+import net.minecraft.client.Minecraft
+
+object LocationUtils {
+    fun playerLocation() = Minecraft.getMinecraft().thePlayer.getModVector()
+}