+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # BedWar Mod - Change Log
 
-# Version 0.1.2
+## Version 0.1.3
+- Made mod detect teammates as eliminated
+- Added some more colours based on health to the team status overlay
+- Allowed the user to customise the delay before the parties stats are sent to chat
+- Made the mod not run when not in a BedWars area
+- Fetch your winstreak from the npc in the hub for the session display
+- Made party match stats work for nicked teammates (but will show them twice)
+- Fixed clicking teammates messages in chat not running `/bws`
+
+## Version 0.1.2
 - Teammate status display - Shows how much health each of your teammates has along with whether they have been eliminated or disconnected
-- BedWars star colour now should reflect Hypixel colours in `/bws`
+- BedWars star colour should now reflect Hypixel colours in `/bws`
 - Made the mod work while in lucky block and allows for repo updates for other weird start messages
 - Various bug fixes
 - Formatting improvements for stats features
 - A few changes that should theoretically improve performance
 
-# Version 0.1.1
+## Version 0.1.1
 - Made all the chat events work with the repo
 - Change some wording of features
 - Tab complete in BedWars games now bases the players off of when the game started